requirement.txt
mlx-lm
langgraph>=0.0.3
langchain>=0.1.0
pydantic>=2.0.0
asyncio>=3.4.3
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dateutil>=2.8.2